Seeder correction
